requirements.txt
asgiref==3.7.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.1
django-import-export==3.3.6
et-xmlfile==1.1.0
MarkupPy==1.14
numpy==1.26.3
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.2.0
psycopg2==2.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.9.0
tzdata==2023.4
xlrd==2.0.1
xlwt==1.3.0
